+++ CHANGES.rst
@@ -8,6 +8,7 @@ Here you can see the full list of changes between each SQLAlchemy-Utils release.
 ^^^^^^^^^^^^^^^^^^^
 
 - Added JSONType
+- NumberRangeType now supports coercing of integer values
 
 
 0.19.0 (2013-10-24)